+++ library/src/main/java/io/sfrei/tracksearch/clients/youtube/YouTubeUtility.java
@@ -111,9 +111,10 @@ class YouTubeUtility {
                     .getFirstField()
                     .get("appendContinuationItemsAction", "continuationItems")
                     .getIndex(1)
+                    .get("continuationItemRenderer", "continuationEndpoint", "continuationCommand")
                     .orElseGet(() -> defaultElement
-                            .getIndex(1))
-                    .get(continuationItemRenderer)
+                            .firstElementFor("continuationItemRenderer")
+                            .get("continuationEndpoint", "continuationCommand"))
                     .getAsString("token");
         }