+++ library/src/test/java/io/sfrei/tracksearch/clients/ClientTestImpl.java
@@ -5,138 +5,100 @@ import io.sfrei.tracksearch.exceptions.TrackSearchException;
 import io.sfrei.tracksearch.tracks.Track;
 import io.sfrei.tracksearch.tracks.TrackList;
 import io.sfrei.tracksearch.tracks.metadata.TrackMetadata;
-import lombok.RequiredArgsConstructor;
+import lombok.Getter;
 import org.junit.jupiter.api.*;
-import org.slf4j.Logger;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 
 import java.net.CookiePolicy;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+@Getter
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public abstract class ClientTestImpl<T extends Track> extends Client implements ClientTest {
+public abstract class ClientTestImpl<T extends Track> extends Client implements ClientTest<T> {
 
     protected final TrackSearchClient<T> searchClient;
 
     private final List<String> searchKeys;
 
-    private final List<TrackList<T>> tracksForSearch;
+    private final List<Named<TrackList<T>>> tracksForSearch;
 
-    protected final Logger logger;
+    private Stream<Named<T>> getAllTracksFromTrackLists() {
+        return tracksForSearch.stream()
+                .flatMap(trackList -> trackList.getPayload().getTracks().stream())
+                .map(track -> Named.of(String.format("%s - %s", track.getTitle(), track.getUrl()), track));
+    }
 
-    public ClientTestImpl(TrackSearchClient<T> trackSearchClient, List<String> searchKeys, Logger logger) {
+    public ClientTestImpl(TrackSearchClient<T> trackSearchClient, List<String> searchKeys) {
         super(CookiePolicy.ACCEPT_ALL, null);
         searchClient = trackSearchClient;
-        logger.debug("Initialized test suite for: {}", searchClient.getClass().getSimpleName());
         this.searchKeys = searchKeys;
-        this.logger = logger;
         tracksForSearch = new ArrayList<>();
     }
 
-    @Override
     @Order(1)
-    @Test
-    public void tracksFoSearch() throws TrackSearchException {
-        logger.info("Search for {}", Arrays.toString(searchKeys.toArray()));
-        final TestLogger testLogger = TestLogger.withLogger(logger);
-        for (String key : searchKeys) {
-            testLogger.logBefore();
+    @Override
+    @ParameterizedTest
+    @MethodSource("getSearchKeys")
+    public void tracksForSearch(String key) throws TrackSearchException {
+        TrackList<T> tracksForSearch = searchClient.getTracksForSearch(key);
+        assertFalse(tracksForSearch.isEmpty());
+        this.tracksForSearch.add(Named.of(tracksForSearch.getQueryInformation().get(TrackList.QUERY_PARAM), tracksForSearch));
 
-            logger.info("Search for: {}", key);
-            TrackList<T> tracksForSearch = searchClient.getTracksForSearch(key);
-            logger.info("Found tracks: {}", tracksForSearch.getTracks().size());
-            assertFalse(tracksForSearch.isEmpty());
-            this.tracksForSearch.add(tracksForSearch);
-        }
     }
 
-    @Override
     @Order(2)
-    @Test
-    public void trackListGotPagingValues() {
-        tracksForSearch.forEach(trackList -> assertTrue(searchClient.hasPagingValues(trackList)));
+    @Override
+    @ParameterizedTest
+    @MethodSource("getTracksForSearch")
+    public void trackListGotPagingValues(TrackList<T> trackList) {
+        assertTrue(searchClient.hasPagingValues(trackList));
     }
 
-    @Override
     @Order(3)
-    @Test
-    public void getNextTracks() throws TrackSearchException {
-        final TestLogger testLogger = TestLogger.withLogger(logger);
-        for (TrackList<T> trackList : tracksForSearch) {
-            testLogger.logBefore();
-
-            logger.info("Next for: {}", trackList.getQueryParam());
-            TrackList<T> nextTracksForSearch = searchClient.getNext(trackList);
-            logger.info("Found tracks: {}", nextTracksForSearch.getTracks().size());
-            assertFalse(nextTracksForSearch.isEmpty());
-        }
+    @Override
+    @ParameterizedTest
+    @MethodSource("getTracksForSearch")
+    public void getNextTracks(TrackList<T> trackList) throws TrackSearchException {
+        TrackList<T> nextTracksForSearch = searchClient.getNext(trackList);
+        assertFalse(nextTracksForSearch.isEmpty());
     }
 
-    @Override
-    @Order(3)
-    @Test
-    public void checkMetadata() {
-        final TestLogger testLogger = TestLogger.withLogger(logger);
-        for (TrackList<T> trackList : tracksForSearch) {
-            for (T track : trackList.getTracks()) {
-                testLogger.logBefore();
-
-                logger.debug("Check track MetaData for: {}", track);
-                assertNotNull(track.getTitle());
-                assertNotNull(track.getLength());
-                assertNotNull(track.getUrl());
-
-                final TrackMetadata trackMetadata = track.getTrackMetadata();
-                assertNotNull(trackMetadata);
-
-                logger.debug("MetaData: {}", trackMetadata);
-                assertNotNull(trackMetadata.getChannelName());
-                assertNotNull(trackMetadata.getChannelUrl());
-                assertNotNull(trackMetadata.getStreamAmount());
-                assertNotNull(trackMetadata.getThumbNailUrl());
-            }
-        }
-    }
-
-    @Override
     @Order(4)
-    @Test
-    public void getStreamUrl() {
-        final TestLogger testLogger = TestLogger.withLogger(logger);
-        for (TrackList<T> trackList : tracksForSearch) {
-            for (T track : trackList.getTracks()) {
-                testLogger.logBefore();
+    @Override
+    @ParameterizedTest
+    @MethodSource("getTracksForSearch")
+    public void checkMetadata(TrackList<T> trackList) {
+        for (T track : trackList.getTracks()) {
 
-                logger.info("Trying to get stream url for: {}", track.toString());
-                String streamUrl = track.getStreamUrl();
-                assertNotNull(streamUrl);
-                logger.info("URL found: {}", streamUrl);
-                final int code = requestAndGetCode(streamUrl);
-                logger.debug("Response code: {}", code);
-                assertNotEquals(FORBIDDEN, code);
-            }
+            assertNotNull(track.getTitle());
+            assertNotNull(track.getLength());
+            assertNotNull(track.getUrl());
+
+            final TrackMetadata trackMetadata = track.getTrackMetadata();
+            assertNotNull(trackMetadata);
+
+            assertNotNull(trackMetadata.getChannelName());
+            assertNotNull(trackMetadata.getChannelUrl());
+            assertNotNull(trackMetadata.getStreamAmount());
+            assertNotNull(trackMetadata.getThumbNailUrl());
         }
     }
 
-    /**
-     * Improve readability, especially in GitHub actions
-     */
-    @RequiredArgsConstructor(staticName = "withLogger")
-    private static class TestLogger {
-
-        private final Logger log;
-
-        private final AtomicInteger pos = new AtomicInteger(1);
-
-        protected void logBefore() {
-            log.info("======= {} =======", String.format("%03d", pos.getAndIncrement()));
-        }
-
+    @Order(5)
+    @Override
+    @ParameterizedTest
+    @MethodSource("getAllTracksFromTrackLists")
+    public void getStreamUrl(Track track) {
+        String streamUrl = track.getStreamUrl();
+        assertNotNull(streamUrl);
+        final int code = requestAndGetCode(streamUrl);
+        assertNotEquals(FORBIDDEN, code);
     }
 
 }