+++ CHANGELOG.md
@@ -4,10 +4,14 @@ Changelog
 0.7.3 - unreleased
 ------------------
 
-- Improved logging
-- Improve tests
-- Avoid parsing advertisements on YouTube search results
+**Features:**
+
 - Additional path for YouTube paging information
+- Improved tests and logging
+
+**Bugfixes:**
+
+- Avoid parsing advertisements on YouTube search results
 
 0.7.2
 -----